pyproject.toml
[project]
name = "bakes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "certifi>=2025.10.5",
    "fastapi>=0.120.2",
    "loguru>=0.7.3",
    "mysql-connector-python>=9.5.0",
    "pymongo>=4.15.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
